Created utils that parser uses with save_persons() and need_update()
